Migrate CheckAndRaise to the C dispatch registry
Make CheckAndRaise a plain Op and register CheckAndRaiseImpl, moving its
c_code and ParamsType into the detached impl.
