+++ pytensor/link/c/dispatch/__init__.py
@@ -1,4 +1,7 @@
 # isort: off
 from pytensor.link.c.dispatch.basic import c_funcify
 
+# Load dispatch specializations
+import pytensor.link.c.dispatch.raise_op
+
 # isort: on
